suggestion

+++ tasks/all/solovyev_d_shell_sort_simple/src/ops_all.cpp
@@ -29,13 +29,13 @@ void solovyev_d_shell_sort_simple_all::TaskALL::ShellSort(std::vector<int>& data
   std::vector<std::thread> threads(num_threads_);
   for (int t = 0; t < num_threads_; ++t) {
     threads[t] = std::thread([&, t] {
-      for (int gap = static_cast<int>(data.size()) / 2; gap > 0; gap /= 2) {
+      for (size_t gap = data.size() / 2; gap > 0; gap /= 2) {
         sync_point.arrive_and_wait();
         for (int i = t; i < gap; i += num_threads_) {
           for (size_t f = i + gap; f < data.size(); f += gap) {
             int val = data[f];
             size_t j = f;
-            while (j >= static_cast<size_t>(gap) && data[j - gap] > val) {
+            while (j >= gap && data[j - gap] > val) {
               data[j] = data[j - gap];
               j -= gap;
             }
